Removed all code from deprecated `people` app
