+++ project_tier/network/models.py
@@ -13,6 +13,7 @@ from wagtail.admin.edit_handlers import (
     FieldPanel, MultiFieldPanel
 )
 from wagtail.images.edit_handlers import ImageChooserPanel
+from wagtail.snippets.models import register_snippet
 
 
 class Person(models.Model):
@@ -234,3 +235,22 @@ class NetworkIndexPage(Page):
 
     class Meta:
         verbose_name = 'Network List Page'
+
+
+@register_snippet
+class PersonCategory(models.Model):
+    title = models.CharField(max_length=255)
+    order = models.IntegerField(
+        null=True, blank=True,
+        help_text="Lower numbers are shown first. If left blank, it will "
+                  "show up at the end."
+    )
+
+    def __str__(self):              # __unicode__ on Python 2
+        # We're returning the string that populates the snippets screen. Note it returns as plain-text
+        return self.title
+
+    class Meta:
+        ordering = ['order', 'title']
+        verbose_name = 'People category'
+        verbose_name_plural = 'People categories'
