--- project_tier/people/templates/people/person_index_page.html
@@ -1,71 +1,0 @@
-{% extends "standard_sidebar_layout.html" %}
-{% load wagtailcore_tags wagtailimages_tags people_tags %}
-
-{% block sidebar %}
-  <div data-sticky-container>
-    <div class="sticky" data-sticky data-sticky-on="large" data-margin-top="3" data-top-anchor="on-this-page:top" data-btm-anchor="main-end:bottom">
-      <nav class="sidebar-nav section-nav">
-        <h3 id="on-this-page">In this section</h3>
-        <ul class="vertical menu" data-magellan data-bar-offset="60">
-          <li class="{% if section.category.is_active %}is-active{% endif %}">
-            <a href="#fellows">Fellows</a>
-          </li>
-          {% for section in self.sections %}
-            <li class="{% if section.category.is_active %}is-active{% endif %}">
-              <a href="#{{section.category.title|slugify}}">{{ section.category.title }}</a>
-            </li>
-          {% endfor %}
-        </ul>
-      </nav>
-    </div>
-  </div>
-{% endblock %}
-
-{% block main_content %}
-  {% if self.body %}
-    <div class="row">
-      <div class="body-content">
-        {{ self.body|richtext }}
-      </div>
-    </div>
-  {% endif %}
-
-  {% if self.fellowship_years %}
-    <div class="fellowships">
-      <h1 id="fellows" data-magellan-target="fellows">Fellows</h1>
-
-      {% with year=page.fellowship_years.0.0 fellows=page.fellowship_years.0.1 %}
-        {% include "includes/person_list.html" with group=year|academic_range list=fellows %}
-      {% endwith %}
-
-      {% with past_fellows=page.fellowship_years|slice:"1:" %}
-        {% if past_fellows %}
-          <ul class="accordion" data-accordion data-multi-expand="true" data-allow-all-closed="true">
-            {% for year, fellows in past_fellows %}
-              <li class="accordion-item" data-accordion-item>
-                <a href="#" class="accordion-title">Fellows, {{ year|academic_range }}</a>
-                <div class="accordion-content" data-tab-content>
-                  {% include "includes/person_list.html" with group=year|academic_range list=fellows %}
-                </div>
-              </li>
-            {% endfor %}
-          </ul>
-        {% endif %}
-      {% endwith %}
-    </div>
-  {% endif %}
-
-  {% if self.sections %}
-    <div class="people">
-      <h1>People</h1>
-
-      {% for section in self.sections %}
-        {% with anchor=section.category.title|slugify %}
-          <a id="{{anchor}}" data-magellan-target="{{anchor}}" name="{{anchor}}"></a>
-        {% endwith %}
-        {% include "includes/person_list.html" with group=section.category.title list=section.people %}
-      {% endfor %}
-    </div>
-  {% endif %}
-
-{% endblock %}
