--- project_tier/people/templates/people/person_page.html
@@ -1,74 +1,0 @@
-{% extends "standard_sidebar_layout.html" %}
-{% load navigation_tags wagtailcore_tags wagtailimages_tags %}
-
-{% block page_banner %}
-  <header class="page-header">
-    <div class="row">
-      <div class="columns">
-        {% breadcrumbs %}
-        {% block page_title %}
-          <h1>{{self.title}}{% if self.academic_title %}<span class="academic-title">, {{ self.academic_title }}</span>{% endif %}</h1>
-        {% endblock %}
-          {% if self.main_job_title %}
-            <div class="job-titles">{{ self.main_job_title }}</div>
-      {% endif %}
-    </div>
-          </div>
-  </header>
-{% endblock %}
-
-
-{% block main_content %}
-
-  <div class="row">
-    <div class="body-content small-12 medium-7 columns">
-        {% if self.introductory_headline %}
-        <p class="overview"><strong>{{ self.introductory_headline }}</strong></p>
-        {% endif %}
-          {% if self.biography %}
-            {{ self.biography|richtext }}
-        {% endif %}
-    </div>
-
-    <aside class="people meta small-12 medium-4 columns">
-        {% if self.image %}
-          {# Person image - first store image as a variable in order to construct an image tag with a class applied #}
-            {% image self.image fill-250x250-c100 as photo %}
-            <img class="img person-img" src="{{ photo.url }}" width="{{ photo.width }}" height="{{ photo.height }}" alt="{{ photo.alt }}" class="img-thumbnail" />
-          {% else %}
-            <img class="placeholder img person-img">
-          {% endif %}
-          {% if self.image_credit %}
-          <figcaption class="people"><span>Photo courtesy of</span> {{ self.image_credit }}</figcaption>
-          {% endif %}
-        <dl class="contact-info">
-            {% if self.email %}
-              <dt>Email</dt>
-              <dd><a href="mailto:{{self.email}}">{{self.email}}</a></dd>
-            {% endif %}
-
-            {% if self.website %}
-              <dt>Website</dt>
-              <dd><a target="_blank" href="{{ self.website }}">{{self.website}}</a></dd>
-            {% endif %}
-
-            {% if self.phone %}
-              <dt>Phone number</dt>
-              <dd>{{ self.phone}}</dd>
-            {% endif %}
-
-            {% if self.location %}
-              <dt>Address</dt>
-              <dd>{{ self.location}}</dd>
-            {% endif %}
-
-          </dl>
-      </aside>
-
-        </div>
-
-      </div>
-    </div>
-  </div>
-  {% include "includes/related_links.html" with related_links=self.related_links.all only %}
-{% endblock %}
